pkg/ttl: move manager job adapter tests to integrationtest
